Rewrite console font rendering
In the code taken from the example of the old bootloader (bootboot), we
were linking against a binary object file of the font and using this to
render the characters. The new implementation simplifies a lot of the
old code by just using a library to parse the embedded raw version of
the font, which is included by the include_bytes!() macro.