Implement really basic heap allocation for kernel
Using the linked_list_allocator crate, we are setting up some really
basic heap allocation for our kernel with this patch.
It allows us to use dynamically allocated data structures like Vec and
also allows us to `Box` values.

To achieve that, we take the first usable memory region that we can
find and use that to allocate physical frame on. Then we take 1M of
memory and map it as the kernel heap and use that in the `LockedHeap`
global allocator.