Refactor console to be a really basic terminal
The console supports user input and actually draws its cursor to the
screen. It also can redraw the screen in its entirety because the
characters that are being put onto the screen are saved inside of a
character buffer.

This makes it pretty straight-forward for future patches to implement
things like backspace, handling commands and all that is needed for a
REPL/shell.