Implement double buffering for the Framebuffer
The framebuffer has drawn everything instantly to the front buffer
before.
That lead to some flickering here and there if we were typing very fast.
This patch implements double buffering, so all the painting happens on a
non-visible back buffer first and the `Framebuffer` struct now provides
a `flip()` function which makes the back buffer visible. Looking pretty
flicker-free now!