Separate terminal from console
To have a better separation of concerns, the input handling now happens
in a dedicated "terminal" module, while the actual handling of a
text-based console stays in the console module.

The terminal module will be expanded with functionality like an I/O loop
(REPL) and all of that stuff in some following commits.