Make the terminal look like a terminal
This patch makes sure that the terminal can print a prompt and get the
user input that comes after the prompt.
It also assures that you can't delete the prompt. Currently, to get what
the user typed, it goes from the offset of the prompt to the cursor.
That means if you press enter while the cursor isn't at the end of the
input, you won't get the complete command. But in general, it works.