Make an "echo" shell builtin
This required a lot of work. I actually implemented syscalls during
making this feature because we needed a way for the stdlib to call into
the kernel to print something to the console on the screen.

We use interrupt number 0x80 just like Linux does and we offer one
syscall that prints a given string to the standard out.