Refactor the architecture
This patch moves the shell into a separate crate to prepare for the
moment that it will become a real program that's being loaded by the
init process.

While doing that, the print() syscall also got expanded to support
another argument which lets us define the channel that we want to print
to (Serial or stdout).